requirements.txt
# Core dependencies
pennylane>=0.30.0
torch>=2.0.0
numpy>=1.20.0
matplotlib>=3.5.0
tqdm>=4.62.0

# Multimodal support dependencies
pillow>=9.0.0  # For image processing
librosa>=0.9.0  # For audio processing
scikit-image>=0.19.0  # For image transformations
torchvision>=0.15.0  # For image models and transforms

# Optional dependencies
pytest>=7.0.0  # For testing
sphinx>=4.3.0  # For documentation
black>=22.3.0  # For code formatting
pylint>=2.12.0  # For linting
jupyter>=1.0.0  # For notebooks